move player stats to own hok
